- Renamed database to dynamoDB - Added some error catching

+++ backend/shared-lib/src/services/favoriteService.ts
@@ -1,11 +1,19 @@
 import {Favorite} from "../model/favorite";
-import {favoriteRepository} from "../repository";
+import {favoriteRepository, statusRepository} from "../repository";
 import {StatusDto} from "../dto/statusDto";
+import {Status} from "../model/status";
 
 
 export class FavoriteService {
 
     async addFavorite(accountID: string, statusID: string) : Promise<Favorite> {
+
+        const status : Status = await statusRepository.getStatusById(statusID);
+
+        if (!status) {
+            return null;
+        }
+
         return await favoriteRepository.persist({
             objectName: "Favorite",
             pkey: accountID,
@@ -14,6 +22,13 @@ export class FavoriteService {
     }
 
     async removeFavorite(accountID: string, statusID: string) : Promise<boolean> {
+
+        const status : Status = await statusRepository.getStatusById(statusID);
+
+        if (!status) {
+            return false;
+        }
+
         return await favoriteRepository.delete(accountID, statusID);
     }
